1.阿博茨科技_CRF表格区域定位项目/Python/requirements.txt
absl-py==0.1.10
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
bleach==1.5.0
certifi==2018.1.18
chardet==3.0.4
decorator==4.2.1
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
grpcio==1.10.0
hb-config==0.5.1
html5lib==0.9999999
idna==2.6
ipython==5.5.0
ipython-genutils==0.2.0
Markdown==2.6.11
mock==2.0.0
numpy==1.14.1
pandas==0.22.0
pathlib2==2.3.0
pbr==3.1.1
pexpect==4.4.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
protobuf==3.5.1
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
requests==2.18.4
scandir==1.7
simplegeneric==0.8.1
six==1.11.0
tensorflow==1.5.0
tensorflow-serving-api==1.5.0
tensorflow-tensorboard==1.5.1
tqdm==4.19.6
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
